Fully move the guts of array_view to span.
Now array_view is a synonym for span, rather than the other way around.
Both will work for legacy client apps for some time, but for OIIO internals
we should start favoring span when convenient.

A few minor fixups to make some existing methods and in-class types
match the C++ standard. One change, as dictated by C++20 std::span<>
specification, is that span.size() returns a ptrdiff_t (signed) rather
than a size_t.  This may create some minor build issues with signed vs
unsigned warnings. Something to watch out for.

+++ src/libOpenImageIO/deepdata.cpp
@@ -527,7 +527,7 @@ DeepData::set_samples (int pixel, int samps)
 void
 DeepData::set_all_samples (array_view<const unsigned int> samples)
 {
-    if (samples.size() != size_t(m_npixels))
+    if (samples.size() != m_npixels)
         return;
     ASSERT (m_impl);
     if (m_impl->m_allocated) {
